semi-final production release
