+++ public/index.html
@@ -1,6 +1,8 @@
 <!DOCTYPE html>
 <html lang="en">
   <head>
+    <link rel="manifest" href="public/manifest.json">
+    <link rel="shortcut icon" href="%PUBLIC_URL%/favicon.ico">
   </head>
   <body>
     <div id="root"></div>
