--- src/components/Frame/Frame.scss
@@ -1,24 +1,0 @@
-#image{
-	  position: absolute;
-    top: 0.1em;
-    left:0.2em;
-    margin: auto;
-    opacity: 0;
-    animation: changeOpacity 3s;
-    animation-fill-mode: forwards;
-}
-
-.frame-Pos {
-  position: relative;
-  top:4em;
-  left: 6em;
-}
-
-@keyframes changeOpacity {
-  0% {
-    opacity: 0;
-  }
-  100% {
-    opacity: 1;
-  }
-}
