+++ tools/buildHTML.js
@@ -7,7 +7,7 @@ fs.readFile('public/index.html', 'utf8', (err, markup) => {
 	}
 
 	const $ = cheerio.load(markup);
-	$('head').prepend('<link rel="stylesheet" href="styles.css">');
+  $('head').prepend('<link rel="stylesheet" href="/styles.css">');
 
 	fs.writeFile('dist/index.html', $.html(), 'utf8', function (err) {
 		if (err){
